Refactored Event Feed
